* updated to v0_4_10 * code cleanup * fixed log with 2048 symbols segments * updated logs with errormessages

+++ source/opentm2/core/utilities/EQFMORPH.cpp
@@ -1401,7 +1401,7 @@ PMORPHHASH MorphHashCreate
   /********************************************************************/
   if ( fOK )
   {
-    pMorphHash->ulMaxElements = min( ulMaxElements,
+    pMorphHash->ulMaxElements = min((unsigned long) ulMaxElements,
                                      (MAX_SEG_SIZE / sizeof(HASHENTRY)) );
     pMorphHash->ulElements  = 0;
   } /* endif */
